[NewOptimizer] Hook up printing, print linetable
This hooks up printing for the new IR and teaches it to print the
linetable. In printing a linetable, I tried to strike a balance
between verboseness (I think the old printing and our printing of LLVM IR
is significantly too verbose). There's a description of the idea in
show.jl