Move the definitions of eltype() and length() in the iteration demo (#27154)
Moves the definitions of eltype() and length() in the basic iteration demo closer to their occurrence in the text. Currently, the methods are defined in code a few paragraphs before they're mentioned in the text, which makes the example a bit harder to follow.

This also helps reinforce the story that iterate() is the *only* required method for iteration, while eltype() and length() are optional improvements.