In union splitting also check type of first argument
I had mistakenly assumed that when we got to unionsplitting the first
type could not be one of the union types we were splitting on. However,
this assumption is incorrect for parameterized types (e.g.
Union{Type{Float64}, Type{BigFloat}} as in #27204). The fix is relatively
simple and consists of also checking the function argument during union
splitting.

Fixes #27204.