Fix reinterpret(T, ::Array) performance
When I originally wrote the new ReinterpretArray code, I made sure that LLVM was able
to optimize reinterpret(::Array) back to a single memory access with appropriate TBAA
and alignment info. Somewhere along the line LLVM lost that ability. While we should
try to recover that capability in LLVM, that showed that that is a relatively brittle
optimization for a very simple operation. So this patch takes a different approach:
We add two new intrinsics `tbaa_pointerref` and `tbaa_pointerset` that behave like
their non-TBAA variants, but additionally take a type to use as the TBAA tag. This
allows us to write a special case for `reinterpret(T, ::Array)` that directly emits
the correct pointer access. It's also a model for what a post-1.0 pure Julia
implementation of `Array` (e.g. on top of a buffer type) may look like.

Fixes #25014