Make Sys.which return nothing instead of throwing
Currently `Sys.which("prog")` will throw an error if `prog` is not
found or is found but is not executable. This changes the function to
instead return `nothing` in those cases, thereby simplifying logic for
situations such as hierarchically choosing an executable based on which
in a list is present.

Fixes #27284.