Docs: manually apply hygiene to [] syntax
Macro hygiene fails to do this automatically,
so we should do it manually (until fixed in lowering)
But first, we have to actually make macro hygiene work for `docm` at all,
so alter the `Core.@doc` to make that possible.

Also apply `@nospecialize` to many of our macro arguments,
in the hopes that this may help reduce code load times.

And try to synchronize our 5 different definitions of `isexpr`
to be merely 3 copies of the same version (in Core.CoreDocs, Core.Compiler, and Base.Meta).