BLAS: avoid calling dlclose
We do not want libblas to get unloaded again here,
but we call this early enough, this is potentially
the only reference to it. That may lead to race
conditions where we try to teardown the library
while it is trying to spawn threads
(spawning threads is not generally permitted on Windows,
but is down anyways by libopenblas.dll).

So instead, we just cache our single reference to it,
and thereby avoid the need to call `dlclose`
to maintain the proper refcount.