Allow `jl_set_const` to overwrite const global (with warning).
Before this commit, `jl_set_const` (and `jl_set_global`) silently do
nothing if the binding is a `const` global variable. This does not match
the `julia` behavior, which is to allow redefining the const global
(with a Warning) as long as *the types are the same*.

This commit changes those two functions to have the same behavior, by
simply delegating to `jl_checked_assignment`.