Fix inlining of exception handlers (#27502)
Usually exception handlers immediately disqualify a function for inlining,
which is why we did not see this before. However, that can of course be
overriden with `@inline`. In the CFG, we use `0` as a virtual predecessor,
to the catch block, in order to force all values to go through the PhiC
mechanism. We need to avoid trying to rename this virtual predecessor.

Fixes #27456