Add test for #27597
The problem here was that we had an incomplete phi from a multi-branch,
but the logic that fills up the phi with `undef`s for incomplete phis,
only added one entry to the phi node. Rather than the two that LLVM was
expecting. Since the previous commit prevents us from turning conditional
branches into multi-edge phis, this problem does not occur anymore.
With this test, this fixes #27597