Change eigvals(::Number) to return a scalar
This is a proposal for how eigvals could return a scalar when given a
scalar input, in line with what svdvals does. (See issue #27687)

Promotion to float is also skipped. (Just like svdvals, eigvals will now
return an integer for scalar real integer input.) For complex scalar
input, the return type will depend on whether the imaginary component is
zero, as before.

Since Julia allows indexing into, and iterating over, a scalar as if it
were a length-1 vector, this change should have minimal effect on
existing code.