Document multiple return types of redirect_stdout and friends
As noted in #27805, the return type of `redirect_stdout` and other functions
such as `redirect_stdin` is not accurate when the optional input argument is
used. Instead of returning a tuple with the (rd, wr) pipe ends for the new
`stdout`, it instead returns the argument itself.

The new documentation accounts for this behavior.