Turn unspecified defaults in at-kwdef into required kwargs (#27987)
The existence of at-kwdef predates the existence of required keyword
arguments; before that, all keyword arguments needed a default value.
So currently when no value is specified for a field in the type
definition given to kwdef, it's assigned a default value based on the
field's type.

With this change, any field not explicitly given a default becomes a
required keyword argument in the resulting type constructor when using
kwdef. In doing this, we're also able to drop the requirement that the
type's fields be typed, and that the type has fields at all.