Fix processed_idx passed to process_newnode!
When inserting new nodes during compaction, we need to rename all arguments.
For performance, we keep track of which nodes we have already processed,
renaming them immediately, and only scheduling those we have not for later
re-processing. However, the was an off-by-one error in this logic such that a phi
node refering to the statement it is being attached to would not get scheduled
for later fixup.