Disallow double edges in IR
A double edge is an edge of the form:
```
4 - goto 5 if not %%x
5 - ...
```
i.e. where both the fall through and the explicit edge of
the conditional branch go to the same basic block (note
however that such branches are basically no-ops). LLVM
allows this (with some restrictions - e.g. if there is
a phi node the value on both edges has to be the same),
so I also decided to allow it. However, it turns out to
cause an enormous amount of pain because edges are no
longer uniquely identified by just the two basic blocks
at the edge's end points. In the near future, I want to
start doing CFG transforms at the IR level and these double
edges require a special case for basically every operation.
Thus, in preparation for that, let's disallow such edges as
an IR invariant. The trade off here is that when we modify
conditional branches now, we have to check whether they
would be a noop and delete them in that case. However,
a) we likely want to do that anyway for IR compactness and
b) it seems a lot less fidegedy than dealing with double edge.