show-ir: print the basic-block label when possible
When basic-block information is available, it's preferable that we print that information
when displaying code IR, as it is more readable (and more consistent
between CodeInfo and IRCode form).

Warning: if the basic blocks are not computed correctly,
this might print misleading information. However, this seems unlikely,
since there should be a basic-block start recorded at every label used as a destination.