Use a custom hashing function for remotecall_fetch
Mitigates issue #25220 (but does not fix it).

This implements a custom hashing function for Distributed to use in determining if it should resend values between workers. Previously, this just used `Base.hash`. This has the possibility that a custom type could opt for a cheaper "partial" or constant-valued hash function, leading to failure to update values between workers. Instead, this implements a generic hashing scheme that works directly on the internal fields --- we don't need to worry about hash equality between different types for this use-case.