Fix arrayref nothrow predicate for undef-able arrays
Previously the nothrow predicate said we were allowed to remove an unused
call if it was marked as inbounds. However, this is only true if none of
the entries of the array can be `#undef` (i.e. if the element type is a
bitstype of bitsunion). Correct the predicate and add a test case.

Fixes #28326