Update types.md
The language was used that type is instantiated to create an instance.   It was explained that Julia instances are different than, say, objects in C++.  In some places the language of instances was used.  In some cases the language of objects.  I brought this up on the discussion forum, and suggested seeing what this would look like if the language of instances was used consistently.  This version uses that language consistently.