Remove excess information from the opening Julia banner
After discussion with Jeff and Viral, the tagline "A fresh approach to technical computing" seems to have become somewhat stale.  We additionally decided that it was unnecessary to show the build timestamp in the initial banner, as well as `Sys.MACHINE` as both are available within `versioninfo()` which is what we want to encourage users to report anyway.  Jeff wanted to keep the build date itself though, and as that's more minimal I think it's a good compromise.