Update variables after normalizing varargs
This code was a bit odd. It started by normalzing the varargs, but
then used `va0` and `va1` from the original, unnormalized type.
Either the unormalized version was intended in which case normalizing
first was wasteful, or the normalized version was intended in which
case the code was incorrct. Talking with Jeff, he believes the latter
was intended, so update the code to do that instead.