Integer type generic splitrange
This came up in https://ci.appveyor.com/project/ChrisRackauckas/lightgraphs-jl/build/1.0.5/job/380b7juluhx3v0jx#L113 where it was seen that using 64-bit integers in an application led to in an inability to use `@distributed` since implicitly assumed that the range type would be `Int`. This allows for a type promotion to occur so that non-`Int` `Integer` collections can be used for the looping range of an `@distributed` for.