Fix out of bounds write in array deserialization
The array was allocated based on the serialized `elsize` of the array,
however, unions get an extra selector array after the regular storage
which was not allocated (because we didn't know it was gonna be a union
array at the time when we allocated it). According to
a48eeefd6207090898f5b1877684aa10f217800a we cannot look at the element
type to allocate the array, so we need to serialize a bit to indicate
that we will have a union array.

Fixes #28998

(cherry picked from commit e7d7259310ca0dd9c89103bb489a7d6034e39460)