Improved English and egs
The examples starting around line 80 showing how to create arrays with dimensions are unhelpful because in every case the dimensions are 2x2 so people don't get any sense of which dimension is which. So I've changed them all to 2x3 which doesn't take up any more vertical space but makes it much clearer which dimension is which.

I don't like the Comprehensions example because the expression is much too complicated: the idea is to show comprehensions, so I think the expression should be a lot simpler.

I changed the `searchsorted` eg to make the searched for number different from all the others to improve clarity.

I think the warning about CartesianIndexes and end should give some hint as to what would happen: would Julia notice and react or not notice and just produce wrong results?

The Iteraration section should explain or at least cross-ref. to how to enumerate, i.e., how to get (index, value) tuples.

Personally I'd use "etc." rather than "etcetera" but that would be a global style change.

I think it'd be a lot easier if you kept .md doc files to <500 lines.