English fixes
Didn't like the 'tandf' examples, in particular f(2) which I felt was misleading. Why have an argument at all? Why not, say, f() = (println("F"); false) and t() = (println("T"); true)?

It seems bad practice to me to use a global in the while loop example. Surely it would be better to create a local scope (e.g., a function)?

I think that the last for loop example should be followed by another one that shows what the text says about using two `for` keywords to make the difference between `for i = 1:2, j = 3:4` and `for i = 1:2 for j = 3:4` clearer.

I don't understand why the InvalidStateException isn't a cross-ref. when all the others listed are?

When describing how to throw an exception the example given throws a built-in DomainError. I would contend that application code should only ever throw custom exceptions so that users can see immediately whether their code or Julia's code has thrown.

PS When editing I have respected the 92 char line width when I've made changes. But I haven't applied it to lines I haven't changed to avoid making the real changes hard to see. I think you'd need a separate pass on the docs to just do the line width with no other changes if you wanted to enforce it since it seems widely ignored in the .md files.