Minor English fixes + 1 semantic change
Most of the info. seems to refer to Unix-like systems. Surely Windows users ought to be better catered for, esp. regarding paths?

I have no idea what the word 'decimation' means; at least not in this context. (I only know it to mean 'wipe-out' or 'destroy'.)

There's no `open` command on Windows that I know of, so I changed that to `start`.

The info. on JULIA_LLVM_ARGS seems incomplete.