More robust show for invalid CFG
Despite the fact that the compiler should never generate invalid
data structures, it is a good practice for the show functions to
be able to handle these and indicate what the problem is. In this
case, if the IR has more instructions than its supposed to according
to the CFG, put all trailing instructions into a "!" basic block to
indicate an error.