Minor English fixes/improvements
Some .md files have two spaces after a period and others have one; oughtn't there to be a standard? (I prefer one; two is from the days of manual typewriters.)

The `threeint()` and `threefloat()` examples are confusing because they have explicit typing which is effectively ignored -- and nothing to do with the point of having them. Why not something like this:
```julia
function addthree(x)
    n = x + 3
    n
end
# and
function addthree(x)
    n = x + 3
end
```