Minor English fixes/improvements
I think that in a comment on one of my earlier fixes someone said that the convention is to put a space around binary operators e.g., `x + y` rather than `x+y`; but this is not done in the `hypot()` function example (and many of the others too). Also, in many -- but by no means all -- examples arguments are given with no space after the comma. I think it is good to show this once (to show that the space isn't needed), but then be consistent thereafter and have the space for clarity's sake.

Maybe Julia ought to have a style-guide specifically for docs? It wouldn't have to be much, the max. recommended line length, whether to write "e.g." or "e.g.," and "etc." or "etcetera" (since all four are used currently), and some preferred syntaxes for examples regarding whitespace.

The explanation of anonymous functions mentions their use for "delaying" a computation; shouldn't it also mention their ability to capture state (i.e., that they can be closures) too?

I really don't like the `f(x,y) = 3x + 4y` vectorization example. I'd prefer an example that readers could work out in their heads (even if roughly) to get a better idea of what's going on. Something like `f(x,y) = x + y`, or at worst `f(x,y) = x + 2y`.