Improved English and egs (#29281)
Improved English and egs

The examples starting around line 80 showing how to create arrays with dimensions are unhelpful because in every case the dimensions are 2x2 so people don't get any sense of which dimension is which. So I've changed them all to 2x3 which doesn't take up any more vertical space but makes it much clearer which dimension is which.

I don't like the Comprehensions example because the expression is much too complicated: the idea is to show comprehensions, so I think the expression should be a lot simpler.

I changed the `searchsorted` eg to make the searched for number different from all the others to improve clarity.

(cherry picked from commit 29b780e09520369b0e62720e5ed7eb7906b57190)