task: drop parent and module references
The parent link can unnecessarily hold memory references live, and the current-module function is gone.
Their uses have been deprecated in favor of the macros @`__FILE__` and @`__MODULE__`.
The `workspace()` function is also gone now, so we can remove the support code for it.

To implement detection of broken incremental compilation (and help
ensure the module is permanently rooted as is generally expected),
keep track of all "open" modules in a hashtable (jl_current_modules).
We can further refine this later, if desired (e.g. also allow any
submodule and of any item in `jl_module_init_order`).

Also removes the global `module_stack` state. This makes us a bit more
robust to async module definition. We could perhaps do even more here,
if necessary (e.g. by checking first that all parents are initialized,
instead of just the direct ancestor). But this is hopefully a vanishingly
rare situation anyways.