Still set rpath-link in RPATH_ESCAPED_ORIGIN
We use this setting for linking curl, which itself needs to find libssh2
and mbedtls. However, while curl will indeed eventually be in the same
directory as these two libraries, it is not during linking, causing link
failures (unless we happen to have global copies of the same libraries
that the linker can pick up). Add an rpath-link command with the absolute
path to our library directory to make sure the dependent libraries get
picked up at link time.