Workaround gcc-5 miscompilation with asserts and double return functions
gcc-5 appears to miscompile jl_type_infer under special circumstances,
leading to segfaults but gcc-7 and clang do fine on this.  The "special
circumstances" are that assertions are enabled when preprocessing
jl_svecref for use within jl_type_infer specifically.

There are various ways to remove the segfaults and force a correct
compilation:

* The given rearrangement

* Replacing jl_svecref in the last part of jl_type_infer with an
  equivalent function, but without the assertions.

* Removing the JL_TRY / JL_CATCH further up.

It's suspicious that the error only occurs when both the double-return
__sigsetjmp() and the no-return __assert_fail() are present inside the
function in sufficient quantities. Perhaps the compiler is failing to
analyze the interaction between these.