Take 2: use symlinks
Instead of transforming the file names, just create a temporary symlink
in the source directory that points back to the root of the source
directory. As long as symlinks work properly, this should be a portable
solution.

(cherry picked from commit 498ff154524065a9cd4a4fb57cedaeef3c5d3c75)