inference: remove ability to infer `(svec(a, b, c)...)`
It seems unlikely that this code pattern is used anywhere, and it is the last instance of this inference hack
we have remaining (after the PartialTuple work). If some package does end up needing it,
we can reimplement it (via adding PartialSimpleVector), but I strongly suspect this is effectively dead-code,
and thus not worthwhile to maintain.