Fix backtrace line numbers following macros
The `CodeInfo` block has a field `codelocs` that indexes into an array
with full `LineInfoNode` objects. This data model allows de-duplication,
and in practice it does (only) `uniq`-style deduplication, i.e. only
consecutive equal objects are deduplicated by this scheme.

The lambda `compact-ir` generates this structure. This commit fixes a bug
where it mixes monotonously increasing the indices (for `codeloc`) with
push/pop operations. This is wrong because the latest `LineInfoNode`
object is always at the end of the current size of the array; not
one position after the current `LineInfoNode`.