Fix deadlock caused by REPL blinking
The REPL beeping (blinking) code made the assumption that the prompt
prefix is always a String. That's not necessarily the case, and in
fact it can be an arbitrary funcion to be called when the prompt is
printed (e.g. TerminalExtensions uses this for the iTerm integration).
However, even worse, this code acquired a lock that it never released
on error, causing a deadlock. Fix the original assumption and also
add a try/catch/finally block such that future errors print the error
rather than deadlocking the terminal.

Fixes https://github.com/Keno/TerminalExtensions.jl/issues/40

(cherry picked from commit 038d8665afd2eeeaca81ba5ebd3da1ece40e69bc)