[AllocOpt] Track deleted instructions
It was possible for the instruction contained in first_safepoint to
be delted, causing use-after-free problems later. Use a WeakVH
rather than a raw Instruction* to null-out the entry when the
Instruction gets deleted. In the future we may want to pay closer
attention to live ranges to avoid this problem entirely.

Fixes #30116

(cherry picked from commit 437d87cfbee3cc976613ab9acc5d96c5166e1072)