Make compilecache atomic
Previously, the cache path is directly passed to the --output-ji
option of the subprocess for generating compile cache.  However, when
several processes are compiling the same package, one process may
modify the file while another process is calculating the checksum,
yielding a broken cache.  This patch let each process create a cache
in a private temporary path and then atomically rename it to the final
cache path.