Use a tier system for platform support
This makes the level to which we support various platforms more specific
and adds a bit of formalism that's typical of other projects, e.g. Rust.
The tier system is explained in the document, but to recap:

1. Builds and passes tests, have binaries and CI
2. Builds, may pass tests, may have binaries
3. May build
4. Hardcore broken

All of our listed platforms have been assigned a tier, which is pretty
obvious for most of them.

I've also added an entry in the support table for Linux with musl libc.
Julia builds and passes most tests on Alpine, but that isn't verified
regularly so I've assigned that tier 3 for now.

[ci skip]