Add `QueuePool` abstraction to `Distributed`
QueuePool

A pool of workers that asynchronously execute jobs passed to them, with
facilities for both in-order and out-of-order result fetching.  This
construct is intended to act like an asynchronous `pmap()`, allowing
streaming parallel operations to dynamically add jobs to a work queue
and fetch results with a minimum of fuss. To use, construct a
`QueuePool`, submit work items with `push_job!()`, and get output
through `fetch_result()`.