Add target defines for wasm/emscripten
This is perhaps a bit of an overeager commit, given that most of the
code to support wasm is not yet upstream, but I think it's important
to get our terminology clear early. This attempts to define platform
defines in both our C code and at the Julia level. For background,
wasm is a general standard for secure execution of untrusted code
arising out of the Javascript world (many web browsers these days
ship wasm support in their Javascript engine). However, wasm is not
tied to JS, so it would be good to avoid hardcoding any such
assumptions into our platform defines (e.g. people are considering
wasm for kernel space vms, etc.). Emscripten on the other hand is
the original C-to-JS compiler, as well as corresponding runtime
libraries, simulatenously fulfulling the role of C library and
operating system kernel (for some features at least - e.g.
providing file system emulation). As such, my suggestion is the
following:

- Sys.OS gets set to `Emscripten`, which is also what the emulated
  `uname` syscall returns when emulated by Emscripten, so is consistent
  with our documentation for what that global means.
- A new `isjsvm` function can be used to determine if we're running
  inside a JavaScript VM (I'd expect this check to be quite common,
  e.g. to call out to the native RegEx engine).
- `isunix` is true for Emscripten, but I don't envision it being true
  generally for either `isjvm` platforms or wasm platforms.
- I didn't currently add a specific check for wasm, since I'd suspect
  this to be rather rare (e.g. only if you wanted to do something like
  emit inline wasm, or have a specific implementation for wasm targets).
  However, you can of course always check Sys.MACHINE for
  wasm32-*-* or wasm64-*-*. If it turns out to be useful, we can add it.

The C code follows the same reasoning and defines _OS_EMSCRIPTEN_ and
_CPU_WASM_. I didn't bother defining something for jsvm, since the changes
to the C code will have to be quite closely tied to specifics of
Emscripten.